refactor: move parser classes to parser package
